requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter
django-debug-toolbar
drf-spectacular
python-dotenv
whitenoise
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
pytest-factoryboy==2.6.0